C++: Split stl.cpp into string.cpp and vector.cpp.
