+++ cpp/ql/test/library-tests/dataflow/taint-tests/taint.expected
@@ -32,37 +32,34 @@
 | movableclass.cpp:55:8:55:9 | s2 | movableclass.cpp:52:23:52:28 | call to source |
 | movableclass.cpp:64:8:64:9 | s2 | movableclass.cpp:23:55:23:60 | call to source |
 | movableclass.cpp:65:11:65:11 | call to operator= | movableclass.cpp:65:13:65:18 | call to source |
-| stl.cpp:113:7:113:7 | a | stl.cpp:109:12:109:17 | call to source |
-| stl.cpp:115:7:115:7 | c | stl.cpp:111:16:111:21 | call to source |
-| stl.cpp:117:9:117:13 | call to c_str | stl.cpp:111:16:111:21 | call to source |
-| stl.cpp:167:13:167:17 | call to c_str | stl.cpp:159:10:159:15 | call to source |
-| stl.cpp:171:13:171:17 | call to c_str | stl.cpp:159:10:159:15 | call to source |
-| stl.cpp:174:13:174:17 | call to c_str | stl.cpp:159:10:159:15 | call to source |
-| stl.cpp:184:7:184:8 | cs | stl.cpp:179:19:179:24 | call to source |
-| stl.cpp:185:7:185:8 | ss | stl.cpp:179:19:179:24 | call to source |
-| stl.cpp:198:7:198:8 | cs | stl.cpp:190:19:190:24 | call to source |
-| stl.cpp:199:7:199:8 | ss | stl.cpp:190:19:190:24 | call to source |
-| stl.cpp:221:8:221:9 | s1 | stl.cpp:216:18:216:23 | call to source |
-| stl.cpp:222:8:222:9 | s2 | stl.cpp:217:20:217:25 | call to source |
-| stl.cpp:223:8:223:9 | s3 | stl.cpp:219:8:219:13 | call to source |
-| stl.cpp:242:8:242:9 | s1 | stl.cpp:238:32:238:37 | call to source |
-| stl.cpp:243:8:243:9 | s2 | stl.cpp:240:20:240:25 | call to source |
-| stl.cpp:252:8:252:8 | c | stl.cpp:250:16:250:21 | call to source |
-| stl.cpp:260:8:260:8 | c | stl.cpp:250:16:250:21 | call to source |
-| stl.cpp:265:8:265:8 | c | stl.cpp:263:28:263:33 | call to source |
-| stl.cpp:307:8:307:8 | x | stl.cpp:300:43:300:49 | source1 |
-| stl.cpp:315:8:315:8 | x | stl.cpp:300:43:300:49 | source1 |
-| stl.cpp:320:8:320:8 | x | stl.cpp:300:43:300:49 | source1 |
-| stl.cpp:335:11:335:11 | call to operator+ | stl.cpp:332:18:332:23 | call to source |
-| stl.cpp:336:11:336:11 | call to operator+ | stl.cpp:332:18:332:23 | call to source |
-| stl.cpp:337:11:337:11 | call to operator+ | stl.cpp:332:18:332:23 | call to source |
-| stl.cpp:340:11:340:11 | call to operator+ | stl.cpp:340:13:340:18 | call to source |
-| stl.cpp:349:8:349:9 | s5 | stl.cpp:345:18:345:23 | call to source |
-| stl.cpp:353:8:353:9 | s6 | stl.cpp:345:18:345:23 | call to source |
-| stl.cpp:358:8:358:9 | s7 | stl.cpp:356:9:356:14 | call to source |
-| stl.cpp:362:8:362:9 | s8 | stl.cpp:345:18:345:23 | call to source |
-| stl.cpp:367:8:367:9 | s9 | stl.cpp:365:13:365:18 | call to source |
-| stl.cpp:375:8:375:10 | s10 | stl.cpp:372:12:372:26 | call to source |
+| string.cpp:15:7:15:7 | a | string.cpp:11:12:11:17 | call to source |
+| string.cpp:17:7:17:7 | c | string.cpp:13:16:13:21 | call to source |
+| string.cpp:19:9:19:13 | call to c_str | string.cpp:13:16:13:21 | call to source |
+| string.cpp:69:13:69:17 | call to c_str | string.cpp:61:10:61:15 | call to source |
+| string.cpp:73:13:73:17 | call to c_str | string.cpp:61:10:61:15 | call to source |
+| string.cpp:76:13:76:17 | call to c_str | string.cpp:61:10:61:15 | call to source |
+| string.cpp:86:7:86:8 | cs | string.cpp:81:19:81:24 | call to source |
+| string.cpp:87:7:87:8 | ss | string.cpp:81:19:81:24 | call to source |
+| string.cpp:100:7:100:8 | cs | string.cpp:92:19:92:24 | call to source |
+| string.cpp:101:7:101:8 | ss | string.cpp:92:19:92:24 | call to source |
+| string.cpp:123:8:123:9 | s1 | string.cpp:118:18:118:23 | call to source |
+| string.cpp:124:8:124:9 | s2 | string.cpp:119:20:119:25 | call to source |
+| string.cpp:125:8:125:9 | s3 | string.cpp:121:8:121:13 | call to source |
+| string.cpp:144:8:144:9 | s1 | string.cpp:140:32:140:37 | call to source |
+| string.cpp:145:8:145:9 | s2 | string.cpp:142:20:142:25 | call to source |
+| string.cpp:154:8:154:8 | c | string.cpp:152:16:152:21 | call to source |
+| string.cpp:162:8:162:8 | c | string.cpp:152:16:152:21 | call to source |
+| string.cpp:167:8:167:8 | c | string.cpp:165:28:165:33 | call to source |
+| string.cpp:182:11:182:11 | call to operator+ | string.cpp:179:18:179:23 | call to source |
+| string.cpp:183:11:183:11 | call to operator+ | string.cpp:179:18:179:23 | call to source |
+| string.cpp:184:11:184:11 | call to operator+ | string.cpp:179:18:179:23 | call to source |
+| string.cpp:187:11:187:11 | call to operator+ | string.cpp:187:13:187:18 | call to source |
+| string.cpp:196:8:196:9 | s5 | string.cpp:192:18:192:23 | call to source |
+| string.cpp:200:8:200:9 | s6 | string.cpp:192:18:192:23 | call to source |
+| string.cpp:205:8:205:9 | s7 | string.cpp:203:9:203:14 | call to source |
+| string.cpp:209:8:209:9 | s8 | string.cpp:192:18:192:23 | call to source |
+| string.cpp:214:8:214:9 | s9 | string.cpp:212:13:212:18 | call to source |
+| string.cpp:222:8:222:10 | s10 | string.cpp:219:12:219:26 | call to source |
 | structlikeclass.cpp:35:8:35:9 | s1 | structlikeclass.cpp:29:22:29:27 | call to source |
 | structlikeclass.cpp:36:8:36:9 | s2 | structlikeclass.cpp:30:24:30:29 | call to source |
 | structlikeclass.cpp:37:8:37:9 | s3 | structlikeclass.cpp:29:22:29:27 | call to source |
@@ -169,3 +166,6 @@
 | taint.cpp:470:7:470:7 | x | taint.cpp:462:6:462:11 | call to source |
 | taint.cpp:471:7:471:7 | y | taint.cpp:462:6:462:11 | call to source |
 | taint.cpp:485:7:485:10 | line | taint.cpp:480:26:480:32 | source1 |
+| vector.cpp:15:8:15:8 | x | vector.cpp:8:43:8:49 | source1 |
+| vector.cpp:23:8:23:8 | x | vector.cpp:8:43:8:49 | source1 |
+| vector.cpp:28:8:28:8 | x | vector.cpp:8:43:8:49 | source1 |
