+++ cpp/ql/test/library-tests/dataflow/taint-tests/test_diff.expected
@@ -30,36 +30,33 @@
 | movableclass.cpp:55:8:55:9 | movableclass.cpp:52:23:52:28 | AST only |
 | movableclass.cpp:64:8:64:9 | movableclass.cpp:23:55:23:60 | AST only |
 | movableclass.cpp:65:11:65:11 | movableclass.cpp:65:13:65:18 | AST only |
-| stl.cpp:115:7:115:7 | stl.cpp:111:16:111:21 | AST only |
-| stl.cpp:117:9:117:13 | stl.cpp:111:16:111:21 | AST only |
-| stl.cpp:167:13:167:17 | stl.cpp:159:10:159:15 | AST only |
-| stl.cpp:171:13:171:17 | stl.cpp:159:10:159:15 | AST only |
-| stl.cpp:174:13:174:17 | stl.cpp:159:10:159:15 | AST only |
-| stl.cpp:184:7:184:8 | stl.cpp:179:19:179:26 | IR only |
-| stl.cpp:185:7:185:8 | stl.cpp:179:19:179:24 | AST only |
-| stl.cpp:198:7:198:8 | stl.cpp:190:19:190:24 | AST only |
-| stl.cpp:199:7:199:8 | stl.cpp:190:19:190:24 | AST only |
-| stl.cpp:221:8:221:9 | stl.cpp:216:18:216:23 | AST only |
-| stl.cpp:222:8:222:9 | stl.cpp:217:20:217:25 | AST only |
-| stl.cpp:223:8:223:9 | stl.cpp:219:8:219:13 | AST only |
-| stl.cpp:242:8:242:9 | stl.cpp:238:32:238:37 | AST only |
-| stl.cpp:243:8:243:9 | stl.cpp:240:20:240:25 | AST only |
-| stl.cpp:252:8:252:8 | stl.cpp:250:16:250:21 | AST only |
-| stl.cpp:260:8:260:8 | stl.cpp:250:16:250:21 | AST only |
-| stl.cpp:265:8:265:8 | stl.cpp:263:28:263:33 | AST only |
-| stl.cpp:307:8:307:8 | stl.cpp:300:43:300:49 | AST only |
-| stl.cpp:315:8:315:8 | stl.cpp:300:43:300:49 | AST only |
-| stl.cpp:320:8:320:8 | stl.cpp:300:43:300:49 | AST only |
-| stl.cpp:335:11:335:11 | stl.cpp:332:18:332:23 | AST only |
-| stl.cpp:336:11:336:11 | stl.cpp:332:18:332:23 | AST only |
-| stl.cpp:337:11:337:11 | stl.cpp:332:18:332:23 | AST only |
-| stl.cpp:340:11:340:11 | stl.cpp:340:13:340:18 | AST only |
-| stl.cpp:349:8:349:9 | stl.cpp:345:18:345:23 | AST only |
-| stl.cpp:353:8:353:9 | stl.cpp:345:18:345:23 | AST only |
-| stl.cpp:358:8:358:9 | stl.cpp:356:9:356:14 | AST only |
-| stl.cpp:362:8:362:9 | stl.cpp:345:18:345:23 | AST only |
-| stl.cpp:367:8:367:9 | stl.cpp:365:13:365:18 | AST only |
-| stl.cpp:375:8:375:10 | stl.cpp:372:12:372:26 | AST only |
+| string.cpp:17:7:17:7 | string.cpp:13:16:13:21 | AST only |
+| string.cpp:19:9:19:13 | string.cpp:13:16:13:21 | AST only |
+| string.cpp:69:13:69:17 | string.cpp:61:10:61:15 | AST only |
+| string.cpp:73:13:73:17 | string.cpp:61:10:61:15 | AST only |
+| string.cpp:76:13:76:17 | string.cpp:61:10:61:15 | AST only |
+| string.cpp:86:7:86:8 | string.cpp:81:19:81:26 | IR only |
+| string.cpp:87:7:87:8 | string.cpp:81:19:81:24 | AST only |
+| string.cpp:100:7:100:8 | string.cpp:92:19:92:24 | AST only |
+| string.cpp:101:7:101:8 | string.cpp:92:19:92:24 | AST only |
+| string.cpp:123:8:123:9 | string.cpp:118:18:118:23 | AST only |
+| string.cpp:124:8:124:9 | string.cpp:119:20:119:25 | AST only |
+| string.cpp:125:8:125:9 | string.cpp:121:8:121:13 | AST only |
+| string.cpp:144:8:144:9 | string.cpp:140:32:140:37 | AST only |
+| string.cpp:145:8:145:9 | string.cpp:142:20:142:25 | AST only |
+| string.cpp:154:8:154:8 | string.cpp:152:16:152:21 | AST only |
+| string.cpp:162:8:162:8 | string.cpp:152:16:152:21 | AST only |
+| string.cpp:167:8:167:8 | string.cpp:165:28:165:33 | AST only |
+| string.cpp:182:11:182:11 | string.cpp:179:18:179:23 | AST only |
+| string.cpp:183:11:183:11 | string.cpp:179:18:179:23 | AST only |
+| string.cpp:184:11:184:11 | string.cpp:179:18:179:23 | AST only |
+| string.cpp:187:11:187:11 | string.cpp:187:13:187:18 | AST only |
+| string.cpp:196:8:196:9 | string.cpp:192:18:192:23 | AST only |
+| string.cpp:200:8:200:9 | string.cpp:192:18:192:23 | AST only |
+| string.cpp:205:8:205:9 | string.cpp:203:9:203:14 | AST only |
+| string.cpp:209:8:209:9 | string.cpp:192:18:192:23 | AST only |
+| string.cpp:214:8:214:9 | string.cpp:212:13:212:18 | AST only |
+| string.cpp:222:8:222:10 | string.cpp:219:12:219:26 | AST only |
 | structlikeclass.cpp:35:8:35:9 | structlikeclass.cpp:29:22:29:27 | AST only |
 | structlikeclass.cpp:36:8:36:9 | structlikeclass.cpp:30:24:30:29 | AST only |
 | structlikeclass.cpp:37:8:37:9 | structlikeclass.cpp:29:22:29:27 | AST only |
@@ -104,3 +101,6 @@
 | taint.cpp:446:7:446:7 | taint.cpp:445:14:445:28 | AST only |
 | taint.cpp:447:9:447:17 | taint.cpp:445:14:445:28 | AST only |
 | taint.cpp:471:7:471:7 | taint.cpp:462:6:462:11 | AST only |
+| vector.cpp:15:8:15:8 | vector.cpp:8:43:8:49 | AST only |
+| vector.cpp:23:8:23:8 | vector.cpp:8:43:8:49 | AST only |
+| vector.cpp:28:8:28:8 | vector.cpp:8:43:8:49 | AST only |
