+++ cpp/ql/test/library-tests/dataflow/taint-tests/test_ir.expected
@@ -1,10 +1,10 @@
 | format.cpp:157:7:157:22 | (int)... | format.cpp:147:12:147:25 | call to source |
 | format.cpp:157:7:157:22 | access to array | format.cpp:147:12:147:25 | call to source |
 | format.cpp:158:7:158:27 | ... + ... | format.cpp:148:16:148:30 | call to source |
-| stl.cpp:113:7:113:7 | (const char *)... | stl.cpp:109:12:109:17 | call to source |
-| stl.cpp:113:7:113:7 | a | stl.cpp:109:12:109:17 | call to source |
-| stl.cpp:184:7:184:8 | cs | stl.cpp:179:19:179:24 | call to source |
-| stl.cpp:184:7:184:8 | cs | stl.cpp:179:19:179:26 | (const char *)... |
+| string.cpp:15:7:15:7 | (const char *)... | string.cpp:11:12:11:17 | call to source |
+| string.cpp:15:7:15:7 | a | string.cpp:11:12:11:17 | call to source |
+| string.cpp:86:7:86:8 | cs | string.cpp:81:19:81:24 | call to source |
+| string.cpp:86:7:86:8 | cs | string.cpp:81:19:81:26 | (const char *)... |
 | structlikeclass.cpp:38:8:38:9 | s4 | structlikeclass.cpp:33:8:33:13 | call to source |
 | structlikeclass.cpp:61:8:61:9 | s2 | structlikeclass.cpp:58:24:58:29 | call to source |
 | structlikeclass.cpp:62:8:62:20 | ... = ... | structlikeclass.cpp:62:13:62:18 | call to source |
